feat: Add database module and service
This commit adds a new module called `DatabaseModule` and a corresponding service called `DatabaseService` to the backend codebase. The `DatabaseModule` is responsible for providing the `DatabaseService` as a provider. The `DatabaseService` extends the `PrismaClient` class and implements the `OnModuleInit` interface, allowing it to connect to the database on module initialization. This change sets up the foundation for interacting with the database in the application.