feat: Add @nestjs/mapped-types dependency and create UpdateBoardDto
This commit adds the "@nestjs/mapped-types" dependency to the project's "package.json" file. This package will be used for creating mapped types in the application.

Additionally, this commit introduces a new file called "update-board.dto.ts" in the "src/boards/dto" directory. The "UpdateBoardDto" class is defined in this file, which extends the "PartialType" class from "@nestjs/mapped-types". This DTO class will be used for updating board data.

These changes set up the foundation for using mapped types and implementing board updates in the application.