refactor: Update npm dependencies and add @faker-js/faker package
This commit updates the npm dependencies in the backend package.json file to include the latest stable versions of the packages. It also adds the @faker-js/faker package as a devDependency. The @faker-js/faker package is used for generating fake data for testing and development purposes. This change ensures that the project is using up-to-date dependencies and provides access to the faker library for generating fake data.